Avoid crashing with `OverflowException` in `ReadBytes`
A user reported this crash:
----
UnobservedTaskException:
System.AggregateException: A Task's exception(s) were not observed either by Waiting on the Task or accessing its Exception property. As a result, the unobserved exception was rethrown by the finalizer thread. (One or more errors occurred. (Arithmetic operation resulted in an overflow.)) ---> System.AggregateException: One or more errors occurred. (Arithmetic operation resulted in an overflow.) ---> System.OverflowException: Arithmetic operation resulted in an overflow.
at
read_memory_64_bit.MemoryReaderFromLiveProcess.ReadBytes(UInt64 startAddress, Int32 length)
at
read_memory_64_bit.EveOnline64.<>c__DisplayClass4_0.<EnumeratePossibleAddressesForUIRootObjects>g__ReadMemoryRegionContentAsULongArray|2(ValueTuple 2 memoryRegion)
at
read_memory_64_bit.EveOnline64.<>c__DisplayClass4_0.<<EnumeratePossibleAddressesForUIRootObjects>g__EnumerateCandidatesForPythonTypeObjectTypeInMemoryRegion|8>d.Move Next(0)
at
System.Linq.Parallel.SelectManyQueryOperator 3.SelectManyQueryOperatorEnumerator 1.MoveNext(TOutput& currentElement, Pair 2& currentKey)
at
System.Linq.Parallel.PipelineSpooling Task 2.SpoolingWork() at System.Linq.Parallel.Spooling TaskBase.Work() at System.Linq.Parallel.QueryTask. BaseWork(Object unused)
at
System.Threading.ExecutionContext.RunFromThreadPoolDispatchLoop (Thread threadPoolThread, ExecutionContext executionContext, ContextCallback callback, Object state)
End of stack trace from previous location -
at
System.Threading.ExecutionContext.RunFromThreadPoolDispatchLoop(Thread threadPoolThread, ExecutionContext executionContext, ContextCallback callback, Object state)

+++ implement/read-memory-64-bit/Program.cs
@@ -112,7 +112,7 @@ class Program
 
                     var memoryRegions =
                         processSampleUnpacked.memoryRegions
-                        .Select(memoryRegion => (memoryRegion.baseAddress, length: memoryRegion.content.Value.Length))
+                        .Select(memoryRegion => (memoryRegion.baseAddress, length: (ulong)memoryRegion.content.Value.Length))
                         .ToImmutableList();
 
                     var uiRootCandidatesAddresses =
